배포 관련 문제

+++ src/components/main/post/postdeletebtn.js
@@ -9,6 +9,8 @@ import { useRouter } from "next/navigation";
 export default function PostDeleteBtn({ post }) {
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [deletePost, setDeletePost] = useState();
+  const [isSubmit, setIsSubmit] = useState(false);
+
   const router = useRouter();
   const openModal = (postId) => {
     setDeletePost(postId);
@@ -20,18 +22,22 @@ export default function PostDeleteBtn({ post }) {
   };
 
   const handleDelete = async (postId) => {
+    if (isSubmit) return;
+
+    setIsSubmit(true);
     try {
       const response = await fetch(`/api/postdelete/${postId}`, {
         method: "DELETE",
       });
       if (response.ok) {
-        console.log("삭제 성공");
         router.refresh();
       } else {
         console.error("게시물 삭제 실패");
       }
     } catch (error) {
       console.error(error);
+    } finally {
+      setIsSubmit(false);
     }
     setIsModalOpen(false);
   };
